added course-work directory, and moved relevant files
